% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_lyapunov.R
\name{solve_lyapunov}
\alias{solve_lyapunov}
\title{Solve the Lyapunov Equation}
\usage{
solve_lyapunov(Theta, M)
}
\arguments{
\item{Theta}{A square matrix.}

\item{M}{A square matrix of the same dimension as \code{Theta}.}
}
\value{
A matrix \code{Q} that solves the equation.
}
\description{
This function solves the continuous Lyapunov equation \eqn{\Theta Q + Q \Theta^T = M}.
}
\examples{
Theta <- matrix(c(1, -2, 3, 4), nrow = 2)
M <- matrix(c(1, 0, 0, 1), nrow = 2)
solve_lyapunov(Theta, M)
}
